conda-requirements.txt
black
mypy
numpy
oauthlib
pillow
pylint
requests
scikit-image
scikit-learn
scipy
tensorflow